Separate HasRelayChains and HasRelayPacket into separate modules
