+++ crates/relayer-components-extra/src/batch/traits/channel.rs
@@ -1,7 +1,7 @@
 use ibc_relayer_components::chain::traits::types::chain::HasChainTypes;
 use ibc_relayer_components::core::traits::sync::Async;
+use ibc_relayer_components::relay::traits::chains::HasRelayChains;
 use ibc_relayer_components::relay::traits::target::ChainTarget;
-use ibc_relayer_components::relay::traits::types::HasRelayChains;
 use ibc_relayer_components::runtime::traits::runtime::HasRuntime;
 
 use crate::batch::types::aliases::MessageBatchSender;
