+++ crates/relayer-components-extra/src/builder/impls/batch.rs
@@ -12,8 +12,8 @@ use ibc_relayer_components::builder::types::aliases::{
 use ibc_relayer_components::chain::traits::types::chain_id::HasChainId;
 use ibc_relayer_components::chain::traits::types::ibc::HasIbcChainTypes;
 use ibc_relayer_components::core::traits::error::HasErrorType;
+use ibc_relayer_components::relay::traits::chains::HasRelayChains;
 use ibc_relayer_components::relay::traits::target::{DestinationTarget, SourceTarget};
-use ibc_relayer_components::relay::traits::types::HasRelayChains;
 use ibc_relayer_components::runtime::traits::mutex::{HasMutex, HasRuntimeWithMutex};
 use ibc_relayer_components::runtime::traits::runtime::HasRuntime;
