+++ crates/relayer-components/src/relay/impls/message_senders/chain_sender.rs
@@ -2,9 +2,9 @@ use async_trait::async_trait;
 
 use crate::chain::traits::message_sender::CanSendMessages;
 use crate::chain::traits::types::ibc::HasIbcChainTypes;
+use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::ibc_message_sender::IbcMessageSender;
 use crate::relay::traits::target::ChainTarget;
-use crate::relay::traits::types::HasRelayChains;
 use crate::std_prelude::*;
 
 pub struct SendIbcMessagesToChain;
