+++ crates/relayer-components/src/relay/impls/messages/skip_update_client.rs
@@ -6,10 +6,10 @@ use crate::chain::traits::queries::consensus_state::CanQueryConsensusState;
 use crate::chain::traits::types::consensus_state::HasConsensusStateType;
 use crate::chain::types::aliases::Height;
 use crate::logger::traits::level::HasBaseLogLevels;
+use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::logs::logger::CanLogRelayTarget;
 use crate::relay::traits::messages::update_client::UpdateClientMessageBuilder;
 use crate::relay::traits::target::ChainTarget;
-use crate::relay::traits::types::HasRelayChains;
 use crate::std_prelude::*;
 
 pub struct SkipUpdateClient<InUpdateClient>(PhantomData<InUpdateClient>);
