+++ crates/relayer-components/src/relay/impls/messages/wait_update_client.rs
@@ -6,10 +6,10 @@ use async_trait::async_trait;
 use crate::chain::traits::queries::status::CanQueryChainStatus;
 use crate::chain::traits::types::ibc::HasIbcChainTypes;
 use crate::logger::traits::level::HasBaseLogLevels;
+use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::logs::logger::CanLogRelayTarget;
 use crate::relay::traits::messages::update_client::UpdateClientMessageBuilder;
 use crate::relay::traits::target::ChainTarget;
-use crate::relay::traits::types::HasRelayChains;
 use crate::runtime::traits::runtime::HasRuntime;
 use crate::runtime::traits::sleep::CanSleep;
 use crate::std_prelude::*;
