+++ crates/relayer-components/src/relay/impls/packet_filters/allow_all.rs
@@ -1,7 +1,7 @@
 use async_trait::async_trait;
 
+use crate::relay::traits::packet::HasRelayPacket;
 use crate::relay::traits::packet_filter::PacketFilter;
-use crate::relay::traits::types::HasRelayPacket;
 use crate::std_prelude::*;
 
 pub struct AllowAll;
