+++ crates/relayer-components/src/relay/impls/packet_filters/and.rs
@@ -2,8 +2,8 @@ use core::marker::PhantomData;
 
 use async_trait::async_trait;
 
+use crate::relay::traits::packet::HasRelayPacket;
 use crate::relay::traits::packet_filter::PacketFilter;
-use crate::relay::traits::types::HasRelayPacket;
 use crate::std_prelude::*;
 
 pub struct And<FilterA, FilterB>(pub PhantomData<(FilterA, FilterB)>);
