+++ crates/relayer-components/src/relay/impls/packet_relayers/ack/base_ack_packet.rs
@@ -2,10 +2,10 @@ use async_trait::async_trait;
 
 use crate::chain::traits::message_builders::ack_packet::CanBuildAckPacketMessage;
 use crate::chain::types::aliases::{Height, WriteAcknowledgementEvent};
+use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::ibc_message_sender::{CanSendIbcMessages, IbcMessageSenderExt};
 use crate::relay::traits::packet_relayers::ack_packet::AckPacketRelayer;
 use crate::relay::traits::target::SourceTarget;
-use crate::relay::traits::types::HasRelayChains;
 use crate::relay::types::aliases::Packet;
 use crate::std_prelude::*;
