+++ crates/relayer-components/src/relay/impls/packet_relayers/general/full_relay.rs
@@ -3,13 +3,14 @@ use async_trait::async_trait;
 use crate::chain::traits::queries::status::CanQueryChainStatus;
 use crate::chain::traits::types::ibc_events::write_ack::HasWriteAcknowledgementEvent;
 use crate::chain::traits::types::status::HasChainStatusType;
+use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::logs::logger::CanLogRelay;
 use crate::relay::traits::logs::packet::CanLogRelayPacket;
+use crate::relay::traits::packet::HasRelayPacketFields;
 use crate::relay::traits::packet_relayer::PacketRelayer;
 use crate::relay::traits::packet_relayers::ack_packet::CanRelayAckPacket;
 use crate::relay::traits::packet_relayers::receive_packet::CanRelayReceivePacket;
 use crate::relay::traits::packet_relayers::timeout_unordered_packet::CanRelayTimeoutUnorderedPacket;
-use crate::relay::traits::types::{HasRelayChains, HasRelayPacketFields};
 use crate::relay::types::aliases::Packet;
 use crate::std_prelude::*;
