+++ crates/relayer-components/src/relay/impls/packet_relayers/general/lock.rs
@@ -5,9 +5,9 @@ use async_trait::async_trait;
 use crate::logger::traits::level::HasBaseLogLevels;
 use crate::relay::traits::logs::logger::CanLogRelay;
 use crate::relay::traits::logs::packet::CanLogRelayPacket;
+use crate::relay::traits::packet::HasRelayPacket;
 use crate::relay::traits::packet_relayer::PacketRelayer;
 use crate::relay::traits::packet_relayers::lock::HasPacketLock;
-use crate::relay::traits::types::HasRelayPacket;
 use crate::std_prelude::*;
 
 /**
