+++ crates/relayer-components/src/relay/impls/packet_relayers/timeout_unordered/timeout_unordered_packet.rs
@@ -2,10 +2,10 @@ use async_trait::async_trait;
 
 use crate::chain::traits::message_builders::timeout_unordered_packet::CanBuildTimeoutUnorderedPacketMessage;
 use crate::chain::types::aliases::Height;
+use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::ibc_message_sender::{CanSendIbcMessages, IbcMessageSenderExt};
 use crate::relay::traits::packet_relayers::timeout_unordered_packet::TimeoutUnorderedPacketRelayer;
 use crate::relay::traits::target::SourceTarget;
-use crate::relay::traits::types::HasRelayChains;
 use crate::relay::types::aliases::Packet;
 use crate::std_prelude::*;
