+++ crates/relayer-components/src/relay/traits/messages/timeout_unordered_packet.rs
@@ -1,7 +1,7 @@
 use async_trait::async_trait;
 
 use crate::chain::types::aliases::{Height, Message};
-use crate::relay::traits::types::HasRelayPacket;
+use crate::relay::traits::packet::HasRelayPacket;
 use crate::std_prelude::*;
 
 /// Encapsulates the capability of the implementer to construct a timeout
