+++ crates/relayer-components/src/relay/traits/mod.rs
@@ -1,11 +1,12 @@
 pub mod auto_relayer;
+pub mod chains;
 pub mod event_relayer;
 pub mod ibc_message_sender;
 pub mod logs;
 pub mod messages;
+pub mod packet;
 pub mod packet_filter;
 pub mod packet_relayer;
 pub mod packet_relayers;
 pub mod target;
 pub mod two_way;
-pub mod types;
