+++ crates/relayer-components/src/relay/traits/packet_relayers/ack_packet.rs
@@ -3,7 +3,7 @@ use async_trait::async_trait;
 use crate::chain::traits::types::ibc_events::write_ack::HasWriteAcknowledgementEvent;
 use crate::chain::types::aliases::{Height, WriteAcknowledgementEvent};
 use crate::core::traits::sync::Async;
-use crate::relay::traits::types::HasRelayPacket;
+use crate::relay::traits::packet::HasRelayPacket;
 use crate::std_prelude::*;
 
 #[async_trait]
