+++ crates/relayer-components/src/relay/traits/packet_relayers/lock.rs
@@ -1,6 +1,6 @@
 use async_trait::async_trait;
 
-use crate::relay::traits::types::HasRelayPacket;
+use crate::relay::traits::packet::HasRelayPacket;
 use crate::std_prelude::*;
 
 /**
