+++ crates/relayer-components/src/relay/traits/packet_relayers/timeout_unordered_packet.rs
@@ -2,7 +2,7 @@ use async_trait::async_trait;
 
 use crate::chain::types::aliases::Height;
 use crate::core::traits::sync::Async;
-use crate::relay::traits::types::HasRelayPacket;
+use crate::relay::traits::packet::HasRelayPacket;
 use crate::std_prelude::*;
 
 /// Encapsulates the capability of a relayer to send timeout packets over
